customer-service/requirements.txt
flask>=3.0
mysql-connector-python>=9.0
kafka-python>=2.0.0
